Extract Ingredient list to its own component and add test
