+++ recipes/src/recipes/RecipeDetail.test.js
@@ -1,4 +1,4 @@
-import { act, render, fireEvent, screen } from '@testing-library/react';
+import { render, fireEvent, screen } from '@testing-library/react';
 
 import RecipeDetail from './RecipeDetail';
 import testData from '../../data/testData';
